update swagger conf
