+++ docs/docs.go
@@ -29,3 +29,31 @@
 //
 // swagger:meta
 package docs
+
+// swagger:model CommonSuccess
+type CommonSuccess struct {
+	// Status of the error
+	// in: int64
+	Status int64 `json:"status"`
+	// Message of the error
+	// in: string
+	Message string `json:"message"`
+}
+
+// swagger:model CommonError
+type CommonError struct {
+	// Status of the error
+	// in: int64
+	Status int64 `json:"status"`
+	// Message of the error
+	// in: string
+	Message string `json:"message"`
+}
+
+// ErrHandler returns error message response
+func ErrHandler(errmessage string) *CommonError {
+	errresponse := CommonError{}
+	errresponse.Status = 0
+	errresponse.Message = errmessage
+	return &errresponse
+}
